+++ uCNC/mcus/avr/mcu_avr.c
@@ -856,7 +856,7 @@ char mcu_getc(void)
 }
 
 //RealTime
-void mcu_freq_to_clocks(float frequency, uint16_t *ticks, uint8_t *prescaller)
+void mcu_freq_to_clocks(float frequency, uint16_t *ticks, uint16_t *prescaller)
 {
     if (frequency < F_STEP_MIN)
         frequency = F_STEP_MIN;
@@ -897,7 +897,7 @@ void mcu_freq_to_clocks(float frequency, uint16_t *ticks, uint8_t *prescaller)
 	In Arduino this is done in TIMER1
 	The frequency range is from 4Hz to F_PULSE
 */
-void mcu_start_step_ISR(uint16_t clocks_speed, uint8_t prescaller)
+void mcu_start_step_ISR(uint16_t clocks_speed, uint16_t prescaller)
 {
     //stops timer
     TCCRB = 0;
@@ -919,7 +919,7 @@ void mcu_start_step_ISR(uint16_t clocks_speed, uint8_t prescaller)
 }
 
 // se implementar amass deixo de necessitar de prescaler
-void mcu_change_step_ISR(uint16_t clocks_speed, uint8_t prescaller)
+void mcu_change_step_ISR(uint16_t clocks_speed, uint16_t prescaller)
 {
     //stops timer
     //TCCRB = 0;
